fix(tab5): zone effectId uint8_t→uint16_t truncation
ZoneState.effectId was uint8_t, truncating K1's 16-bit hex effect IDs
(0x0100+) to their low byte. WsMessageRouter parsed effectId via
.as<uint8_t>() compounding the loss. Round-trip effect IDs through
Tab5 produced garbage values, causing "Invalid effectId" errors.

Fix: ZoneState.effectId → uint16_t, WsMessageRouter parses via
int→uint16_t with range validation.

+++ tab5-encoder/src/ui/ZoneComposerUI.h
@@ -23,7 +23,7 @@
  * Zone state for display
  */
 struct ZoneState {
-    uint8_t effectId = 0;
+    uint16_t effectId = 0;
     char effectName[48] = {0};
     uint8_t speed = 25;
     uint8_t paletteId = 0;
